refactor(cron): 将 CronJobDetail 拆分为独立路由 /app/cron/:jobId
Cron 列表页和详情页共用一个页面并以内联状态管理切换，页面逻辑复杂。
将详情页提取为独立路由，列表点击跳转至 /app/cron/:jobId，通过 emitter
事件同步数据变更，降低单文件复杂度。
